Move fitbit routes into a separate file.
